+++ perl/TableSort.pm
@@ -6,8 +6,8 @@
 #===================================================================
 package TableSort;
 
+use LinkedSpec;
 use TableGrep;
-use PPlugin;
 
 
 sub GenericFilter {
@@ -36,7 +36,7 @@ my ($conf, $a2d_ref, $filter_seq) = @_;
 			      my @varargs;
 
 			      push @varargs, $conf, $info, $a2d_ref, $maptable, \@args;
-			      $filteredata = PPlugin->get ("genericfilter_$action")->(@varargs);
+			      $filteredata = LinkedSpec::get_plugin("genericfilter_$action")->(@varargs);
 			      die "$@, " if $@;
 
 			     } else {